refactor: 리뷰어 피드백 반영 - AnnotationHandlerMapping 역할 분리
